Move the confirmation-related structures out of the forum module.
