+++ rs/cli/src/unit_tests/update_unassigned_nodes.rs
@@ -3,8 +3,9 @@ use std::{str::FromStr, sync::Arc};
 use crate::artifact_downloader::MockArtifactDownloader;
 use crate::auth::Neuron;
 use crate::commands::{update_unassigned_nodes::UpdateUnassignedNodes, ExecutableCommand};
+use crate::confirm::ConfirmationModeOptions;
 use crate::cordoned_feature_fetcher::MockCordonedFeatureFetcher;
-use crate::forum::{ConfirmationModeOptions, ForumParameters, SubmissionParameters};
+use crate::forum::{ForumParameters, SubmissionParameters};
 use crate::ic_admin::MockIcAdmin;
 use ic_management_backend::health::MockHealthStatusQuerier;
 use ic_management_backend::{lazy_git::MockLazyGit, lazy_registry::MockLazyRegistry, proposal::MockProposalAgent};
